excercise2: https://ad-teaching.informatik.uni-freiburg.de/route-planning-ss2012/exercise-sheet-2.pdf

+++ src/main/java/graph/Arc.java
@@ -5,5 +5,5 @@ package graph;
  * @param headNodeId (tailNode ------> headNode)
  * @param cost is time taken to travel to this node.
  */
-public record Arc(int headNodeId, long cost) {
+public record Arc(int headNodeId, int cost) {
 }
